\name{SpatialMethodsWorkbench-package}
\alias{SpatialMethodsWorkbench}
\alias{SpatialMethodsWorkbench-package}
\title{Spatial Methods Workbench}
\description{
An R interface to the public analysis workflows used by the Spatial Methods
Workbench Shiny application. The package validates mapped tabular inputs,
runs compatible methods, and creates citation-aware reproducibility bundles.
}
\docType{package}
\author{Sagnik Bhadury}
\seealso{\url{https://github.com/sagnikbhadury/SpatialMethodsWorkbench}}
